backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
aiosqlite==0.20.0
python-multipart==0.0.18
